resolved errors with nodemailer and gmail config and changed sending email to dummy nursery finds email
