+++ src/controllers/auth.js
@@ -1,11 +1,11 @@
 //double check top stuff with team's sections
-
+require('dotenv').config()
 const bcrypt = require('bcryptjs')
-const nodemailer = require('nodemailer')
 const jwt = require('jsonwebtoken')
 const User = require('../../models/User')
 const { StatusCodes } = require('http-status-codes')
 const { BadRequestError, UnauthenticatedError } = require('../errors')
+const transporter = require('../middleware/email')
 
 const register = async (req, res) => {
 	//double check model name
@@ -71,7 +71,7 @@ const resetPassword = async (req, res) => {
 const forgotPassword = async (req, res) => {
 	const { email } = req.body
 	const user = await User.findOne({ email })
-	console.log(user)
+	//console.log(user)
 
 	if (!user) {
 		throw new UnauthenticatedError('User with this email does not exist')
@@ -81,24 +81,13 @@ const forgotPassword = async (req, res) => {
 		expiresIn: '20m',
 	})
 	const link = `http://localhost:8000/api/v1/auth/reset-password/${user.id}/${token}`
-	console.log(link)
+	//console.log(link)
 	res.send('Password reset link has been sent to your email')
 
-	// create reusable transporter object using the default SMTP transport
-	let transporter = nodemailer.createTransport({
-		service: 'gmail',
-		auth: {
-			type: 'OAuth2',
-			user: process.env.MAIL_USERNAME,
-			pass: process.env.MAIL_PASSWORD,
-			clientId: process.env.OAUTH_CLIENTID,
-			clientSecret: process.env.OAUTH_CLIENT_SECRET,
-			refreshToken: process.env.OAUTH_REFRESH_TOKEN,
-		},
-	})
+	
 
 	let mailOptions = {
-		from: '"NurseyFinds" <noreply@gmail.com>', // sender address
+		from: `${process.env.MAIL_USERNAME}@gmail.com`, // sender address
 		to: `${user.email}`, // list of receivers
 		subject: 'Link to Reset Password', // Subject line
 		text: `${link}`, // plain text body
